SQL query generator/requirements.txt
streamlit==1.36.0
google-generativeai